requirements.txt
flask
py2neo
pyltp
bs4
pandas
lxml
faiss-cpu
nltk
transformers
pyaudio
numpy
scipy
matplotlib
python_speech_features
keras
tensorflow
scikit-learn
baidu-aip
librosa
zhconv 
torch
peft
langconv
blingfire
xmltodict
fastNLP
flask_cors
pdf2docx